requirements.txt
numpy
pandas
scikit-learn
matplotlib
dataframe-image
mealpy
PyYAML